requirements.txt
Flask==3.0.3
gunicorn==23.0.0
Pillow==11.2.1
requests==2.32.3
